Add files via upload
This rmd contains all of my code so far, I created a dataframe called Lavaflow2 which has Tggc combined with Tggc-h (as Tggc2) and Tgo with Tgo-h (as Tgo2) and the observations in categories Tgch, Tgmh, Tgtb, Tgtbr are deleted since they contained less than 5 observations and couldn't be combined with any other category. All of the repeated measurements were removed from Lavaflow1 and Lavaflow2.